Add nidx to ndbfixtures.index
